Improve reward readability, optimize code
